% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_fraktionen_strassen}
\alias{abfallnavi_fraktionen_strassen}
\title{List waste fractions for a street}
\usage{
abfallnavi_fraktionen_strassen(strassen_id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{strassen_id}{Street id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with waste fractions.
}
\description{
List waste fractions for a street
}
