% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_hausnummern}
\alias{abfallnavi_hausnummern}
\title{List house numbers for a street}
\usage{
abfallnavi_hausnummern(strassen_id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{strassen_id}{Street id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with house numbers.
}
\description{
List house numbers for a street
}
