% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_kalender_download}
\alias{abfallnavi_kalender_download}
\title{Download calendar file}
\usage{
abfallnavi_kalender_download(
  region,
  format,
  jahr,
  ort,
  strasse,
  hnr,
  fraktion,
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{region}{Region code.}

\item{format}{File format.}

\item{jahr}{Year.}

\item{ort}{Place name.}

\item{strasse}{Street id.}

\item{hnr}{House number id.}

\item{fraktion}{Fraction ids.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with raw file bytes.
}
\description{
Download calendar file
}
\details{
Downloads a calendar file for the requested address and fraction.
}
