% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_ort}
\alias{abfallnavi_ort}
\title{Get a place by id}
\usage{
abfallnavi_ort(ort_id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{ort_id}{Place id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with place metadata.
}
\description{
Get a place by id
}
