% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_strassen}
\alias{abfallnavi_strassen}
\title{List streets for a place}
\usage{
abfallnavi_strassen(ort_id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{ort_id}{Place id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with streets.
}
\description{
List streets for a place
}
