% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_termine_hausnummern}
\alias{abfallnavi_termine_hausnummern}
\title{List collection dates for a house number}
\usage{
abfallnavi_termine_hausnummern(
  hausnummern_id,
  fraktion,
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{hausnummern_id}{House number id.}

\item{fraktion}{Fraction ids.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with collection dates.

Includes \code{date_time} as POSIXct in Europe/Berlin.
}
\description{
List collection dates for a house number
}
