% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{autobahn_roads}
\alias{autobahn_roads}
\title{List Autobahn road ids}
\usage{
autobahn_roads()
}
\value{
A tibble with available road ids.

Use \code{bunddev_parameters("autobahn")} to see the currently valid parameters
if the API has changed.
}
\description{
List Autobahn road ids
}
