% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_auth.R
\name{bunddev_auth_token}
\alias{bunddev_auth_token}
\title{Get raw auth token for an API}
\usage{
bunddev_auth_token(api)
}
\arguments{
\item{api}{Registry id.}
}
\value{
The raw token value, or NULL if no auth configured.
}
\description{
Get raw auth token for an API
}
\keyword{internal}
