% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_charging_station_details}
\alias{bunddev_autobahn_charging_station_details}
\title{Get Autobahn charging station details}
\usage{
bunddev_autobahn_charging_station_details(station_id, flatten = FALSE)
}
\arguments{
\item{station_id}{Charging station identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with charging station details.
}
\description{
Get Autobahn charging station details
}
