% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_charging_stations}
\alias{bunddev_autobahn_charging_stations}
\title{List Autobahn charging stations}
\usage{
bunddev_autobahn_charging_stations(
  road_id,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with charging stations.
}
\description{
List Autobahn charging stations
}
