% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_closures}
\alias{bunddev_autobahn_closures}
\title{List Autobahn closures}
\usage{
bunddev_autobahn_closures(road_id, flatten = FALSE)
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with closures.
}
\description{
List Autobahn closures
}
