% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_parking_lorries}
\alias{bunddev_autobahn_parking_lorries}
\title{List Autobahn lorry parking areas}
\usage{
bunddev_autobahn_parking_lorries(road_id, flatten = FALSE)
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with lorry parking areas.
}
\description{
List Autobahn lorry parking areas
}
