% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_parking_lorry_details}
\alias{bunddev_autobahn_parking_lorry_details}
\title{Get Autobahn lorry parking details}
\usage{
bunddev_autobahn_parking_lorry_details(
  lorry_id,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{lorry_id}{Lorry parking identifier.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with lorry parking details.
}
\description{
Get Autobahn lorry parking details
}
