% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_roads}
\alias{bunddev_autobahn_roads}
\title{List Autobahn road ids}
\usage{
bunddev_autobahn_roads()
}
\value{
A tibble with available road ids.
}
\description{
List Autobahn road ids
}
