% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_autobahn.R
\name{bunddev_autobahn_roadwork_details}
\alias{bunddev_autobahn_roadwork_details}
\title{Get Autobahn roadwork details}
\usage{
bunddev_autobahn_roadwork_details(roadwork_id, flatten = FALSE)
}
\arguments{
\item{roadwork_id}{Roadwork identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with roadwork details.
}
\description{
Get Autobahn roadwork details
}
