% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_roadworks}
\alias{bunddev_autobahn_roadworks}
\title{List Autobahn roadworks}
\usage{
bunddev_autobahn_roadworks(road_id, flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with roadworks.
}
\description{
List Autobahn roadworks
}
