% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_warnings}
\alias{bunddev_autobahn_warnings}
\title{List Autobahn warnings}
\usage{
bunddev_autobahn_warnings(road_id, flatten = FALSE)
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with warnings.
}
\description{
List Autobahn warnings
}
