% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_autobahn.R
\name{bunddev_autobahn_webcam_details}
\alias{bunddev_autobahn_webcam_details}
\title{Get Autobahn webcam details}
\usage{
bunddev_autobahn_webcam_details(webcam_id, flatten = FALSE)
}
\arguments{
\item{webcam_id}{Webcam identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with webcam details.
}
\description{
Get Autobahn webcam details
}
