% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_autobahn.R
\name{bunddev_autobahn_webcams}
\alias{bunddev_autobahn_webcams}
\title{List Autobahn webcams}
\usage{
bunddev_autobahn_webcams(road_id, flatten = FALSE)
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with webcams.
}
\description{
List Autobahn webcams
}
