% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_bewerberboerse.R
\name{bunddev_bewerberboerse_search}
\alias{bunddev_bewerberboerse_search}
\title{Search the Bewerberboerse API}
\usage{
bunddev_bewerberboerse_search(params = list(), flatten = FALSE)
}
\arguments{
\item{params}{List of query parameters.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with tidied results.
}
\description{
Search the Bewerberboerse API
}
