% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_openapi.R
\name{bunddev_endpoints}
\alias{bunddev_endpoints}
\title{List OpenAPI endpoints for an API}
\usage{
bunddev_endpoints(id)
}
\arguments{
\item{id}{Registry id.}

\item{name}{Optional parameter name to filter.}

\item{path}{Optional endpoint path to filter.}

\item{method}{Optional HTTP method to filter.}
}
\value{
A tibble with endpoints.
}
\description{
List OpenAPI endpoints for an API
}
