% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_handelsregister.R
\name{bunddev_handelsregister_search}
\alias{bunddev_handelsregister_search}
\title{Search the Handelsregister portal}
\usage{
bunddev_handelsregister_search(
  schlagwoerter,
  schlagwort_optionen = c("all", "min", "exact"),
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{schlagwoerter}{Search terms.}

\item{schlagwort_optionen}{Keyword options.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with search results.
}
\description{
Search the Handelsregister portal
}
