% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{bunddev_info}
\alias{bunddev_info}
\title{Get a registry entry by id}
\usage{
bunddev_info(id)
}
\arguments{
\item{id}{Registry id.}
}
\value{
A tibble with a single registry entry.
}
\description{
Get a registry entry by id
}
