% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_time.R
\name{bunddev_ms_to_posix}
\alias{bunddev_ms_to_posix}
\title{Convert milliseconds to POSIXct}
\usage{
bunddev_ms_to_posix(value, tz = "Europe/Berlin")
}
\arguments{
\item{value}{Timestamp in milliseconds.}

\item{tz}{Timezone for conversion.}
}
\value{
POSIXct timestamp.
}
\description{
Convert milliseconds to POSIXct
}
\details{
Converts epoch milliseconds to POSIXct in the requested timezone.
}
\examples{
bunddev_ms_to_posix(1704067200000, tz = "Europe/Berlin")

}
\seealso{
Other OpenAPI: 
\code{\link{bunddev_cache_dir}()},
\code{\link{bunddev_call}()},
\code{\link{bunddev_call_tidy}()},
\code{\link{bunddev_endpoints}()},
\code{\link{bunddev_parameter_values}()},
\code{\link{bunddev_parameters}()},
\code{\link{bunddev_parameters_for}()},
\code{\link{bunddev_spec}()},
\code{\link{bunddev_spec_path}()},
\code{\link{bunddev_timestamp_to_ms}()}
}
\concept{OpenAPI}
