% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_auth.R
\name{bunddev_oauth_client_id}
\alias{bunddev_oauth_client_id}
\title{Get OAuth2 client ID for an API}
\usage{
bunddev_oauth_client_id(api)
}
\arguments{
\item{api}{Registry id.}
}
\value{
Client ID string.
}
\description{
Get OAuth2 client ID for an API
}
\details{
Returns the OAuth2 client ID from env var or default config.
}
\keyword{internal}
