% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_openapi.R
\name{bunddev_parameter_values}
\alias{bunddev_parameter_values}
\title{Extract parameter enum values}
\usage{
bunddev_parameter_values(endpoint, name)
}
\arguments{
\item{endpoint}{Adapter function or its name.}

\item{name}{Parameter name.}
}
\value{
A character vector of enum values.
}
\description{
Extract parameter enum values
}
\details{
Returns unique enum values for a parameter defined on the adapter endpoint.
}
\examples{
\dontrun{
bunddev_parameter_values(smard_timeseries, "resolution")
}

}
\seealso{
\code{\link[=bunddev_parameters_for]{bunddev_parameters_for()}} to inspect all parameters for an adapter.
}
