% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_openapi.R
\name{bunddev_parameters_for}
\alias{bunddev_parameters_for}
\title{List OpenAPI parameters for a specific adapter}
\usage{
bunddev_parameters_for(endpoint)
}
\arguments{
\item{endpoint}{Adapter function or its name.}
}
\value{
A tibble with parameter metadata.
}
\description{
List OpenAPI parameters for a specific adapter
}
