% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_registry.R
\name{bunddev_search}
\alias{bunddev_search}
\title{Search registry entries}
\usage{
bunddev_search(q)
}
\arguments{
\item{q}{Search query.}
}
\value{
A tibble of matching registry entries.
}
\description{
Search registry entries
}
