% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{bunddev_spec_path}
\alias{bunddev_spec_path}
\title{Build a cache path for an API spec}
\usage{
bunddev_spec_path(id)
}
\arguments{
\item{id}{Registry id.}
}
\value{
File path where the API specification is cached (character scalar).
}
\description{
Build a cache path for an API spec
}
\details{
Determines the cache file name based on the spec URL extension (.yaml/.json).
}
\examples{
\dontrun{
bunddev_spec_path("smard")
}

}
\seealso{
\code{\link[=bunddev_spec]{bunddev_spec()}} to download and parse specs.

Other OpenAPI: 
\code{\link{bunddev_cache_dir}()},
\code{\link{bunddev_call}()},
\code{\link{bunddev_call_tidy}()},
\code{\link{bunddev_endpoints}()},
\code{\link{bunddev_ms_to_posix}()},
\code{\link{bunddev_parameter_values}()},
\code{\link{bunddev_parameters}()},
\code{\link{bunddev_parameters_for}()},
\code{\link{bunddev_spec}()},
\code{\link{bunddev_timestamp_to_ms}()}
}
\concept{OpenAPI}
