% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_tagesschau.R
\name{bunddev_tagesschau_channels}
\alias{bunddev_tagesschau_channels}
\title{Fetch Tagesschau channels}
\usage{
bunddev_tagesschau_channels(flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with channels.
}
\description{
Fetch Tagesschau channels
}
