% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_bundesrat.R
\name{bundesrat_plenum_naechste_sitzungen}
\alias{bundesrat_plenum_naechste_sitzungen}
\title{Get Bundesrat upcoming plenum sessions}
\usage{
bundesrat_plenum_naechste_sitzungen(
  view = "render[iOSDetailsWithoutInnerDate]",
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{view}{Rendering mode for the XML output.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with upcoming session metadata.
}
\description{
Get Bundesrat upcoming plenum sessions
}
\details{
Returns upcoming Bundesrat sessions. Official docs:
https://bundesrat.api.bund.dev.
}
\examples{
\dontrun{
bundesrat_plenum_naechste_sitzungen()
}

}
