% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_bundestag.R
\name{bundestag_article}
\alias{bundestag_article}
\title{Get a Bundestag news article}
\usage{
bundestag_article(article_id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{article_id}{Article id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with article fields.
}
\description{
Get a Bundestag news article
}
\details{
Returns a single Bundestag news article in XML format.
Official docs: https://bundesapi.github.io/bundestag-api/.
}
\examples{
\dontrun{
bundestag_article(849630)
}

}
