% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_bundestag.R
\name{bundestag_mdb_bio}
\alias{bundestag_mdb_bio}
\title{Get a Bundestag biography}
\usage{
bundestag_mdb_bio(mdb_id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{mdb_id}{Member id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with biography fields.
}
\description{
Get a Bundestag biography
}
\details{
Returns details for a single member of parliament.
Official docs: https://bundesapi.github.io/bundestag-api/.
}
\examples{
\dontrun{
bundestag_mdb_bio(1769)
}

}
