% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_bundestag.R
\name{bundestag_mdb_index}
\alias{bundestag_mdb_index}
\title{List Bundestag members of parliament}
\usage{
bundestag_mdb_index(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with member entries.
}
\description{
List Bundestag members of parliament
}
\details{
Returns the index of members of parliament.
Official docs: https://bundesapi.github.io/bundestag-api/.
}
\examples{
\dontrun{
bundestag_mdb_index()
}

}
