% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_destatis.R
\name{destatis_data_table}
\alias{destatis_data_table}
\title{Retrieve Destatis table data}
\usage{
destatis_data_table(
  name,
  params = list(),
  username = "GAST",
  password = "GAST",
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{name}{Table id.}

\item{params}{Query parameters.}

\item{username}{Genesis username (default "GAST").}

\item{password}{Genesis password (default "GAST").}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with table data in a text column.
}
\description{
Retrieve Destatis table data
}
\details{
Returns table data as text (csv by default).
Official docs: https://github.com/bundesAPI/destatis-api.
}
\examples{
\dontrun{
destatis_data_table("12411-0001")
}

}
