% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_diga.R
\name{diga_charge_item_definitions}
\alias{diga_charge_item_definitions}
\title{List DiGA prescription units}
\usage{
diga_charge_item_definitions(
  params = list(),
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns.}
}
\value{
A tibble containing DiGA prescription unit (ChargeItemDefinition)
resources from the FHIR API. Structure depends on \code{flatten} setting.
}
\description{
List DiGA prescription units
}
