% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dip_bundestag.R
\name{dip_bundestag_aktivitaet}
\alias{dip_bundestag_aktivitaet}
\title{Get a DIP Aktivit채t}
\usage{
dip_bundestag_aktivitaet(
  aktivitaet_id,
  params = list(),
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{aktivitaet_id}{Aktivit채t id.}

\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with Aktivit채t metadata.
}
\description{
Get a DIP Aktivit채t
}
\examples{
\dontrun{
dip_bundestag_aktivitaet(1493545)
}

}
