% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dip_bundestag.R
\name{dip_bundestag_drucksache_text}
\alias{dip_bundestag_drucksache_text}
\title{Get a DIP Drucksache text}
\usage{
dip_bundestag_drucksache_text(
  drucksache_id,
  params = list(),
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{drucksache_id}{Drucksache id.}

\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with Drucksache text metadata.
}
\description{
Get a DIP Drucksache text
}
\examples{
\dontrun{
dip_bundestag_drucksache_text(68852)
}

}
