% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dip_bundestag.R
\name{dip_bundestag_person}
\alias{dip_bundestag_person}
\title{Get a DIP Person}
\usage{
dip_bundestag_person(person_id, params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{person_id}{Person id.}

\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with person metadata.
}
\description{
Get a DIP Person
}
\examples{
\dontrun{
dip_bundestag_person(1728)
}

}
