% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dip_bundestag.R
\name{dip_bundestag_plenarprotokoll}
\alias{dip_bundestag_plenarprotokoll}
\title{Get a DIP Plenarprotokoll}
\usage{
dip_bundestag_plenarprotokoll(
  plenarprotokoll_id,
  params = list(),
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{plenarprotokoll_id}{Plenarprotokoll id.}

\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with Plenarprotokoll metadata.
}
\description{
Get a DIP Plenarprotokoll
}
\examples{
\dontrun{
dip_bundestag_plenarprotokoll(908)
}

}
