% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dip_bundestag.R
\name{dip_bundestag_vorgangsposition}
\alias{dip_bundestag_vorgangsposition}
\title{Get a DIP Vorgangsposition}
\usage{
dip_bundestag_vorgangsposition(
  vorgangsposition_id,
  params = list(),
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{vorgangsposition_id}{Vorgangsposition id.}

\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with Vorgangsposition metadata.
}
\description{
Get a DIP Vorgangsposition
}
\examples{
\dontrun{
dip_bundestag_vorgangsposition(173376)
}

}
