% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dwd.R
\name{dwd_alpine_forecast_text}
\alias{dwd_alpine_forecast_text}
\title{Fetch DWD alpine forecast text}
\usage{
dwd_alpine_forecast_text(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with the forecast text.

Use \code{bunddev_parameters("dwd")} to see the currently valid parameters
if the API has changed.
}
\description{
Fetch DWD alpine forecast text
}
