% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_hilfsmittel.R
\name{hilfsmittel_nachweisschema}
\alias{hilfsmittel_nachweisschema}
\title{Get Hilfsmittel verification schema details}
\usage{
hilfsmittel_nachweisschema(
  id,
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{id}{Nachweisschema id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with Nachweisschema details.
}
\description{
Get Hilfsmittel verification schema details
}
\details{
Returns detail information for a Nachweisschema.
Official docs: https://github.com/bundesAPI/hilfsmittel-api.
}
\examples{
\dontrun{
hilfsmittel_nachweisschema("a3d37017-2c91-4d6d-bbbe-4002d2868044")
}

}
