% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_hilfsmittel.R
\name{hilfsmittel_produkt}
\alias{hilfsmittel_produkt}
\title{Get Hilfsmittel product details}
\usage{
hilfsmittel_produkt(
  id,
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{id}{Produkt id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with product details.
}
\description{
Get Hilfsmittel product details
}
\details{
Returns detail information for a single product.
Official docs: https://github.com/bundesAPI/hilfsmittel-api.
}
\examples{
\dontrun{
hilfsmittel_produkt("f41f52a6-5d2d-4dd3-9d0e-39675ceca7f3")
}

}
