% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_hilfsmittel.R
\name{hilfsmittel_untergruppe}
\alias{hilfsmittel_untergruppe}
\title{Get Hilfsmittel subgroup details}
\usage{
hilfsmittel_untergruppe(
  id,
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{id}{Untergruppe id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with subgroup details.
}
\description{
Get Hilfsmittel subgroup details
}
\details{
Returns details for a subgroup (Untergruppe).
Official docs: https://github.com/bundesAPI/hilfsmittel-api.
}
\examples{
\dontrun{
hilfsmittel_untergruppe("c92d1976-d3cb-4b9f-bcdf-805272a9ea86")
}

}
