% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_hochwasserzentralen.R
\name{hochwasserzentralen_bundeslaender}
\alias{hochwasserzentralen_bundeslaender}
\title{List flood portal states and connected regions}
\usage{
hochwasserzentralen_bundeslaender(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with Bundesland metadata.
}
\description{
List flood portal states and connected regions
}
\details{
Returns metadata for all Bundeslaender and connected regions in the
hochwasserzentralen.de portal. Official docs:
https://bundesapi.github.io/hochwasserzentralen-api/.
}
\examples{
\dontrun{
hochwasserzentralen_bundeslaender()
}

}
