% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_hochwasserzentralen.R
\name{hochwasserzentralen_bundesland_info}
\alias{hochwasserzentralen_bundesland_info}
\title{Get flood portal metadata for a Bundesland}
\usage{
hochwasserzentralen_bundesland_info(
  bundesland_id,
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{bundesland_id}{Bundesland id (e.g., "HE").}

\item{safe}{Logical; if \code{TRUE} (default), apply rate-limiting and cache
GET responses to \code{tools::R_user_dir("bunddev", "cache")}.}

\item{refresh}{Logical; if \code{TRUE}, ignore cached responses and re-fetch
from the API (default \code{FALSE}).}
}
\value{
A tibble with one row of metadata for the selected Bundesland id.
}
\description{
Get flood portal metadata for a Bundesland
}
\details{
Returns metadata for a single Bundesland or region in the
hochwasserzentralen.de portal. API documentation: \url{https://bundesapi.github.io/hochwasserzentralen-api/}.
}
\examples{
\dontrun{
hochwasserzentralen_bundesland_info("HE")
}

}
\seealso{
Other Hochwasserzentralen: 
\code{\link{hochwasserzentralen_bundeslaender}()},
\code{\link{hochwasserzentralen_bundesland_geojson}()},
\code{\link{hochwasserzentralen_lagepegel}()},
\code{\link{hochwasserzentralen_pegel_info}()}
}
\concept{Hochwasserzentralen}
