% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_interpol.R
\name{interpol_un_notices}
\alias{interpol_un_notices}
\title{List Interpol UN notices}
\usage{
interpol_un_notices(
  params = list(),
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with notices metadata.
}
\description{
List Interpol UN notices
}
\details{
Returns UN notices from Interpol. Official docs:
https://interpol.api.bund.dev.
}
\examples{
\dontrun{
interpol_un_notices(params = list(resultPerPage = 10, page = 1))
}

}
