% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_interpol.R
\name{interpol_yellow_notice_images}
\alias{interpol_yellow_notice_images}
\title{List Interpol yellow notice images}
\usage{
interpol_yellow_notice_images(
  notice_id,
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{notice_id}{Notice ID.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with image metadata.

Use \code{bunddev_parameters("interpol")} to see the currently valid parameters
if the API has changed.
}
\description{
List Interpol yellow notice images
}
\details{
Returns images metadata for a yellow notice.
Official docs: https://interpol.api.bund.dev.
}
\examples{
\dontrun{
interpol_yellow_notice_images("2014-5590")
}

}
