% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_jobsuche.R
\name{jobsuche_logo}
\alias{jobsuche_logo}
\title{Fetch Jobsuche employer logo}
\usage{
jobsuche_logo(hash_id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{hash_id}{Logo hash id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with raw logo bytes.

Use \code{bunddev_parameters("jobsuche")} to see the currently valid parameters
if the API has changed.
}
\description{
Fetch Jobsuche employer logo
}
