% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_luftqualitaet.R
\name{luftqualitaet_airquality_limits}
\alias{luftqualitaet_airquality_limits}
\title{Get air quality date limits}
\usage{
luftqualitaet_airquality_limits(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with date limits.
}
\description{
Get air quality date limits
}
\details{
Returns date limits for air quality measurements. Official docs:
https://luftqualitaet.api.bund.dev.
}
