% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_luftqualitaet.R
\name{luftqualitaet_measures_limits}
\alias{luftqualitaet_measures_limits}
\title{Get measurement date limits}
\usage{
luftqualitaet_measures_limits(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with date limits.
}
\description{
Get measurement date limits
}
\details{
Returns date limits for measurement metadata. Official docs:
https://luftqualitaet.api.bund.dev.
}
