% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_luftqualitaet.R
\name{luftqualitaet_meta}
\alias{luftqualitaet_meta}
\title{List combined metadata}
\usage{
luftqualitaet_meta(params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with combined metadata.
}
\description{
List combined metadata
}
\details{
Returns combined metadata for a given use case. Official docs:
https://luftqualitaet.api.bund.dev.
}
