% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_luftqualitaet.R
\name{luftqualitaet_stationsettings}
\alias{luftqualitaet_stationsettings}
\title{List station settings}
\usage{
luftqualitaet_stationsettings(params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with station settings metadata.
}
\description{
List station settings
}
\details{
Returns station settings metadata. Official docs:
https://luftqualitaet.api.bund.dev.
}
