% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_luftqualitaet.R
\name{luftqualitaet_stationtypes}
\alias{luftqualitaet_stationtypes}
\title{List station types}
\usage{
luftqualitaet_stationtypes(params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with station types metadata.
}
\description{
List station types
}
\details{
Returns station types metadata. Official docs:
https://luftqualitaet.api.bund.dev.
}
