% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_luftqualitaet.R
\name{luftqualitaet_transgressions}
\alias{luftqualitaet_transgressions}
\title{List transgressions}
\usage{
luftqualitaet_transgressions(params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with transgressions data.
}
\description{
List transgressions
}
\details{
Returns exceedances (transgressions) data. Official docs:
https://luftqualitaet.api.bund.dev.
}
