% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_luftqualitaet.R
\name{luftqualitaet_transgressiontypes}
\alias{luftqualitaet_transgressiontypes}
\title{List transgression types}
\usage{
luftqualitaet_transgressiontypes(params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with transgression type metadata.
}
\description{
List transgression types
}
\details{
Returns available transgression types. Official docs:
https://luftqualitaet.api.bund.dev.
}
