% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_archive_mowas}
\alias{nina_archive_mowas}
\title{Get MOWAS archive entry}
\usage{
nina_archive_mowas(identifier, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{identifier}{Warning identifier.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with archive entry data.
}
\description{
Get MOWAS archive entry
}
