% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_archive_mowas_mapping}
\alias{nina_archive_mowas_mapping}
\title{Get MOWAS archive mapping}
\usage{
nina_archive_mowas_mapping(identifier, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{identifier}{Warning identifier.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with archive mapping data.
}
\description{
Get MOWAS archive mapping
}
