% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_covid_infos}
\alias{nina_covid_infos}
\title{Get COVID info data}
\usage{
nina_covid_infos(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with COVID info data.
}
\description{
Get COVID info data
}
