% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_covid_map}
\alias{nina_covid_map}
\title{Get COVID map data}
\usage{
nina_covid_map(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with COVID map data.
}
\description{
Get COVID map data
}
