% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_covid_rules}
\alias{nina_covid_rules}
\title{Get COVID rules for an ARS}
\usage{
nina_covid_rules(ars, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{ars}{ARS code.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with COVID rules data.
}
\description{
Get COVID rules for an ARS
}
