% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_covid_ticker}
\alias{nina_covid_ticker}
\title{Get COVID ticker}
\usage{
nina_covid_ticker(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with COVID ticker data.
}
\description{
Get COVID ticker
}
