% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_covid_ticker_message}
\alias{nina_covid_ticker_message}
\title{Get COVID ticker message}
\usage{
nina_covid_ticker_message(id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{id}{Ticker message id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with ticker message data.
}
\description{
Get COVID ticker message
}
