% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_event_code}
\alias{nina_event_code}
\title{Get an event code file}
\usage{
nina_event_code(filename, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{filename}{Event code filename.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with raw event code bytes.
}
\description{
Get an event code file
}
