% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_event_codes}
\alias{nina_event_codes}
\title{List NINA event codes}
\usage{
nina_event_codes(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with event codes.

Includes \code{last_modification_time} as POSIXct in Europe/Berlin.
}
\description{
List NINA event codes
}
