% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_faqs}
\alias{nina_faqs}
\title{List FAQs}
\usage{
nina_faqs(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with FAQs.

Includes \code{last_modification_time} as POSIXct in Europe/Berlin.
}
\description{
List FAQs
}
