% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_logo}
\alias{nina_logo}
\title{Get a logo file}
\usage{
nina_logo(filename, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{filename}{Logo file name.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with raw logo bytes.
}
\description{
Get a logo file
}
