% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_logos}
\alias{nina_logos}
\title{List NINA logos}
\usage{
nina_logos(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with logo metadata.

Includes \code{last_modification_time} as POSIXct in Europe/Berlin.
}
\description{
List NINA logos
}
