% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_mapdata}
\alias{nina_mapdata}
\title{List map data}
\usage{
nina_mapdata(
  source = c("katwarn", "biwapp", "mowas", "dwd", "lhp", "police"),
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{source}{Map data source.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with map data entries.

Includes \code{start_date_time} as POSIXct in Europe/Berlin.
}
\description{
List map data
}
\details{
Valid sources are: \code{katwarn}, \code{biwapp}, \code{mowas}, \code{dwd}, \code{lhp}, \code{police}.
}
