% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_mowas_rss}
\alias{nina_mowas_rss}
\title{Get MOWAS RSS feed}
\usage{
nina_mowas_rss(ars, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{ars}{ARS code.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with RSS XML text.
}
\description{
Get MOWAS RSS feed
}
