% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_notfalltipps}
\alias{nina_notfalltipps}
\title{List emergency tips}
\usage{
nina_notfalltipps(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with emergency tips.

Includes \code{last_modification_time} as POSIXct in Europe/Berlin.
}
\description{
List emergency tips
}
