% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_version}
\alias{nina_version}
\title{Get data version info}
\usage{
nina_version(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with version data.

Includes \code{last_modification_time} as POSIXct in Europe/Berlin.
}
\description{
Get data version info
}
