% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_warning_geojson}
\alias{nina_warning_geojson}
\title{Get a NINA warning (GeoJSON)}
\usage{
nina_warning_geojson(identifier, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{identifier}{Warning identifier.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with geojson payload.
}
\description{
Get a NINA warning (GeoJSON)
}
\details{
Returns a warning in GeoJSON format.
}
