% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_pflanzenschutzmittelzulassung.R
\name{psm_kultur_gruppen}
\alias{psm_kultur_gruppen}
\title{List crop groups}
\usage{
psm_kultur_gruppen(params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{params}{Additional query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with crop group data.
}
\description{
List crop groups
}
\details{
Returns hierarchical crop group classifications.
}
\examples{
\dontrun{
psm_kultur_gruppen()
}

}
