% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_pflanzenschutzmittelzulassung.R
\name{psm_schadorg_gruppen}
\alias{psm_schadorg_gruppen}
\title{List pest groups}
\usage{
psm_schadorg_gruppen(params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{params}{Additional query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with pest group data.
}
\description{
List pest groups
}
\details{
Returns hierarchical pest/pathogen group classifications.
}
\examples{
\dontrun{
psm_schadorg_gruppen()
}

}
