% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_pflanzenschutzmittelzulassung.R
\name{psm_stand}
\alias{psm_stand}
\title{Get data version}
\usage{
psm_stand(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with version information.
}
\description{
Get data version
}
\details{
Returns the release date/version of the plant protection product database.
}
\examples{
\dontrun{
psm_stand()
}

}
