% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_travelwarning.R
\name{travelwarning_representatives_country}
\alias{travelwarning_representatives_country}
\title{List German representatives in foreign countries}
\usage{
travelwarning_representatives_country(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with representatives.

Includes \code{last_modified_time} as POSIXct in Europe/Berlin.
}
\description{
List German representatives in foreign countries
}
\details{
Returns a list of German representatives in foreign countries. Official docs:
https://travelwarning.api.bund.dev.
}
\examples{
\dontrun{
travelwarning_representatives_country()
}

}
