% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_zoll.R
\name{zoll_produktgruppen}
\alias{zoll_produktgruppen}
\title{Get product groups}
\usage{
zoll_produktgruppen(last_modified_date = NULL, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{last_modified_date}{Optional datetime to retrieve only changes after
this date.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with product group data.
}
\description{
Get product groups
}
\details{
Returns product group classifications.
}
\examples{
\dontrun{
zoll_produktgruppen()
}

}
