% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_fraktionen}
\alias{abfallnavi_fraktionen}
\title{List waste fractions}
\usage{
abfallnavi_fraktionen(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with waste fractions.
}
\description{
List waste fractions
}
