% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_fraktionen_hausnummern}
\alias{abfallnavi_fraktionen_hausnummern}
\title{List waste fractions for a house number}
\usage{
abfallnavi_fraktionen_hausnummern(hausnummern_id, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{hausnummern_id}{House number id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with waste fractions.
}
\description{
List waste fractions for a house number
}
