% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_abfallnavi.R
\name{abfallnavi_termine_strassen}
\alias{abfallnavi_termine_strassen}
\title{List collection dates for a street}
\usage{
abfallnavi_termine_strassen(
  strassen_id,
  fraktion,
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{strassen_id}{Street id.}

\item{fraktion}{Fraction ids.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with collection dates.

Includes \code{date_time} as POSIXct in Europe/Berlin.
}
\description{
List collection dates for a street
}
