% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{autobahn_closures}
\alias{autobahn_closures}
\title{List Autobahn closures}
\usage{
autobahn_closures(road_id, flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with closures.

Use \code{bunddev_parameters("autobahn")} to see the currently valid parameters
if the API has changed.
}
\description{
List Autobahn closures
}
