% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{autobahn_parking_lorry_details}
\alias{autobahn_parking_lorry_details}
\title{Get Autobahn lorry parking details}
\usage{
autobahn_parking_lorry_details(
  lorry_id,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{lorry_id}{Lorry parking identifier.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with lorry parking details.

Includes \code{start_time} as POSIXct in Europe/Berlin when available.

Use \code{bunddev_parameters("autobahn")} to see the currently valid parameters
if the API has changed.
}
\description{
Get Autobahn lorry parking details
}
