% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{autobahn_roads}
\alias{autobahn_roads}
\title{List Autobahn road ids}
\usage{
autobahn_roads()
}
\value{
A tibble with available road ids.
}
\description{
List Autobahn road ids
}
\details{
Lists Autobahn road ids from the Autobahn App API (Autobahn GmbH).
Official docs: https://autobahn.api.bund.dev.
}
\examples{
\dontrun{
autobahn_roads()
}
}
\seealso{
\code{\link[=autobahn_roadworks]{autobahn_roadworks()}} and \code{\link[=autobahn_warnings]{autobahn_warnings()}} for road-specific data.
}
