% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{autobahn_warning_details}
\alias{autobahn_warning_details}
\title{Get Autobahn warning details}
\usage{
autobahn_warning_details(warning_id, flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{warning_id}{Warning identifier.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with warning details.

Includes \code{start_time} as POSIXct in Europe/Berlin when available.

Use \code{bunddev_parameters("autobahn")} to see the currently valid parameters
if the API has changed.
}
\description{
Get Autobahn warning details
}
