% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{autobahn_webcam_details}
\alias{autobahn_webcam_details}
\title{Get Autobahn webcam details}
\usage{
autobahn_webcam_details(webcam_id, flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{webcam_id}{Webcam identifier.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with webcam details.
}
\description{
Get Autobahn webcam details
}
