% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_bewerberboerse.R
\name{bewerberboerse_details}
\alias{bewerberboerse_details}
\title{Retrieve Bewerberboerse candidate details}
\usage{
bewerberboerse_details(referenznummer, flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{referenznummer}{Bewerber referenznummer.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.

Use \code{bunddev_parameters("bewerberboerse")} to see the currently valid
parameters if the API has changed.}
}
\value{
A tibble with tidied results.
}
\description{
Retrieve Bewerberboerse candidate details
}
