% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_bewerberboerse.R
\name{bewerberboerse_search}
\alias{bewerberboerse_search}
\title{Search the Bewerberboerse API}
\usage{
bewerberboerse_search(params = list(), flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{params}{List of query parameters.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with tidied results.
}
\description{
Search the Bewerberboerse API
}
