% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_auth.R
\name{bunddev_auth_get}
\alias{bunddev_auth_get}
\title{Get authentication configuration for an API}
\usage{
bunddev_auth_get(api)
}
\arguments{
\item{api}{Registry id.}
}
\value{
A list with auth settings.
}
\description{
Get authentication configuration for an API
}
