% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_auth.R
\name{bunddev_auth_get}
\alias{bunddev_auth_get}
\title{Get authentication configuration for an API}
\usage{
bunddev_auth_get(api)
}
\arguments{
\item{api}{Registry id.}
}
\value{
A named list with auth settings for \code{api}. If no configuration was
set, returns defaults with \code{type = "none"}.
}
\description{
Get authentication configuration for an API
}
\details{
Returns the stored auth configuration for the API or a default \code{none} entry
if no auth has been configured.
}
\examples{
bunddev_auth_get("jobsuche")

}
\seealso{
\code{\link[=bunddev_auth_set]{bunddev_auth_set()}} to configure credentials.

Other Authentication: 
\code{\link{bunddev_auth_set}()}
}
\concept{Authentication}
