% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bunddev_auth_set}
\alias{bunddev_auth_set}
\title{Set authentication configuration for an API}
\usage{
bunddev_auth_set(api, type = "api_key", env_var)
}
\arguments{
\item{api}{Registry id.}

\item{type}{Authentication type.}

\item{env_var}{Environment variable containing credentials.}
}
\value{
The updated auth configuration.
}
\description{
Set authentication configuration for an API
}
