% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_charging_stations}
\alias{bunddev_autobahn_charging_stations}
\title{List Autobahn charging stations}
\usage{
bunddev_autobahn_charging_stations(road_id, flatten = FALSE)
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with charging stations.
}
\description{
List Autobahn charging stations
}
