% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_autobahn.R
\name{bunddev_autobahn_closure_details}
\alias{bunddev_autobahn_closure_details}
\title{Get Autobahn closure details}
\usage{
bunddev_autobahn_closure_details(closure_id, flatten = FALSE)
}
\arguments{
\item{closure_id}{Closure identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with closure details.
}
\description{
Get Autobahn closure details
}
