% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_autobahn.R
\name{bunddev_autobahn_parking_lorry_details}
\alias{bunddev_autobahn_parking_lorry_details}
\title{Get Autobahn lorry parking details}
\usage{
bunddev_autobahn_parking_lorry_details(lorry_id, flatten = FALSE)
}
\arguments{
\item{lorry_id}{Lorry parking identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with lorry parking details.
}
\description{
Get Autobahn lorry parking details
}
