% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_autobahn.R
\name{bunddev_autobahn_roadworks}
\alias{bunddev_autobahn_roadworks}
\title{List Autobahn roadworks}
\usage{
bunddev_autobahn_roadworks(road_id, flatten = FALSE)
}
\arguments{
\item{road_id}{Road identifier.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with roadworks.
}
\description{
List Autobahn roadworks
}
