% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_bewerberboerse.R
\name{bunddev_bewerberboerse_details}
\alias{bunddev_bewerberboerse_details}
\title{Retrieve Bewerberboerse candidate details}
\usage{
bunddev_bewerberboerse_details(referenznummer, flatten = FALSE)
}
\arguments{
\item{referenznummer}{Bewerber referenznummer.}

\item{flatten}{Logical; drop nested list columns.}
}
\value{
A tibble with tidied results.
}
\description{
Retrieve Bewerberboerse candidate details
}
