% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{bunddev_cache_dir}
\alias{bunddev_cache_dir}
\title{Locate the bunddev cache directory}
\usage{
bunddev_cache_dir()
}
\value{
Path to the \code{bunddev} cache directory (character scalar).
}
\description{
Locate the bunddev cache directory
}
\details{
The cache directory is used to store downloaded OpenAPI specs and cached
API responses. Use this to inspect or clean cached files.
}
\examples{
\dontrun{
bunddev_cache_dir()
}

}
\seealso{
\code{\link[=bunddev_spec]{bunddev_spec()}} to download specs, and \code{\link[=bunddev_spec_path]{bunddev_spec_path()}} to locate
a specific spec file.

Other OpenAPI: 
\code{\link{bunddev_call}()},
\code{\link{bunddev_call_tidy}()},
\code{\link{bunddev_endpoints}()},
\code{\link{bunddev_ms_to_posix}()},
\code{\link{bunddev_parameter_values}()},
\code{\link{bunddev_parameters}()},
\code{\link{bunddev_parameters_for}()},
\code{\link{bunddev_spec}()},
\code{\link{bunddev_spec_path}()},
\code{\link{bunddev_timestamp_to_ms}()}
}
\concept{OpenAPI}
