% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_call.R
\name{bunddev_call}
\alias{bunddev_call}
\title{Call an API operation}
\usage{
bunddev_call(
  api,
  operation_id,
  params = list(),
  parse = c("json", "text", "raw"),
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{api}{Registry id.}

\item{operation_id}{OpenAPI operationId.}

\item{params}{List of parameters.}

\item{parse}{Response parsing mode.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached GET responses.}
}
\value{
Parsed response.
}
\description{
Call an API operation
}
