% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_tidy.R
\name{bunddev_call_tidy}
\alias{bunddev_call_tidy}
\title{Call an API operation and return tidy data}
\usage{
bunddev_call_tidy(api, operation_id, params = list(), ...)
}
\arguments{
\item{api}{Registry id.}

\item{operation_id}{OpenAPI operationId.}

\item{params}{List of parameters.}

\item{...}{Additional arguments passed to the tidier.}
}
\value{
A tibble with tidied results.
}
\description{
Call an API operation and return tidy data
}
