% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi.R
\name{bunddev_endpoints}
\alias{bunddev_endpoints}
\title{List OpenAPI endpoints for an API}
\usage{
bunddev_endpoints(id)
}
\arguments{
\item{id}{Registry id.}
}
\value{
A tibble with endpoints.
}
\description{
List OpenAPI endpoints for an API
}
