% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{bunddev_list}
\alias{bunddev_list}
\title{List registry entries}
\usage{
bunddev_list(tag = NULL, auth = NULL)
}
\arguments{
\item{tag}{Optional tag to filter on.}

\item{auth}{Optional auth type to filter on.}
}
\value{
A tibble of registry entries.
}
\description{
List registry entries
}
