% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_time.R
\name{bunddev_ms_to_posix}
\alias{bunddev_ms_to_posix}
\title{Convert milliseconds to POSIXct}
\usage{
bunddev_ms_to_posix(value, tz = "Europe/Berlin")
}
\arguments{
\item{value}{Timestamp in milliseconds.}

\item{tz}{Timezone for conversion.}
}
\value{
POSIXct timestamp.
}
\description{
Convert milliseconds to POSIXct
}
\details{
Converts epoch milliseconds to POSIXct in the requested timezone.
}
\examples{
bunddev_ms_to_posix(1704067200000, tz = "Europe/Berlin")

}
