% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_openapi.R
\name{bunddev_parameters}
\alias{bunddev_parameters}
\title{List OpenAPI parameters for an API}
\usage{
bunddev_parameters(id, name = NULL, path = NULL, method = NULL)
}
\arguments{
\item{id}{Registry id.}
}
\value{
A tibble with parameter metadata.
}
\description{
List OpenAPI parameters for an API
}
