% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_throttle.R
\name{bunddev_rate_limit_get}
\alias{bunddev_rate_limit_get}
\title{Get API rate limit configuration}
\usage{
bunddev_rate_limit_get(api)
}
\arguments{
\item{api}{Registry id.}
}
\value{
The rate limit configuration.
}
\description{
Get API rate limit configuration
}
