% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throttle.R
\name{bunddev_rate_limit_set}
\alias{bunddev_rate_limit_set}
\title{Set API rate limit configuration}
\usage{
bunddev_rate_limit_set(api, max_per_hour)
}
\arguments{
\item{api}{Registry id.}

\item{max_per_hour}{Maximum number of calls per hour.}
}
\value{
The stored rate limit configuration.
}
\description{
Set API rate limit configuration
}
