% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{bunddev_spec}
\alias{bunddev_spec}
\title{Retrieve a cached API spec}
\usage{
bunddev_spec(id, refresh = FALSE)
}
\arguments{
\item{id}{Registry id.}

\item{refresh}{Logical; refresh cached spec.}
}
\value{
Parsed OpenAPI spec.
}
\description{
Retrieve a cached API spec
}
