% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{bunddev_spec_path}
\alias{bunddev_spec_path}
\title{Build a cache path for an API spec}
\usage{
bunddev_spec_path(id)
}
\arguments{
\item{id}{Registry id.}
}
\value{
File path for the cached spec.
}
\description{
Build a cache path for an API spec
}
\details{
Determines the cache file name based on the spec URL extension (.yaml/.json).
}
\examples{
\dontrun{
bunddev_spec_path("smard")
}

}
\seealso{
\code{\link[=bunddev_spec]{bunddev_spec()}} to download and parse specs.
}
