% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{bunddev_spec_path}
\alias{bunddev_spec_path}
\title{Build a cache path for an API spec}
\usage{
bunddev_spec_path(id)
}
\arguments{
\item{id}{Registry id.}
}
\value{
File path for the cached spec.
}
\description{
Build a cache path for an API spec
}
