% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_time.R
\name{bunddev_timestamp_to_ms}
\alias{bunddev_timestamp_to_ms}
\title{Convert timestamps to milliseconds}
\usage{
bunddev_timestamp_to_ms(value, tz = "Europe/Berlin")
}
\arguments{
\item{value}{Timestamp as numeric (ms), Date, or POSIXct.}

\item{tz}{Timezone for Date/POSIXct conversion.}
}
\value{
Numeric timestamp in milliseconds.
}
\description{
Convert timestamps to milliseconds
}
\details{
This helper standardizes timestamps for APIs that expect epoch milliseconds.
Dates are interpreted in the supplied timezone.
}
\examples{
bunddev_timestamp_to_ms(as.POSIXct("2024-01-01 00:00:00", tz = "Europe/Berlin"))

}
\seealso{
Other OpenAPI: 
\code{\link{bunddev_cache_dir}()},
\code{\link{bunddev_call}()},
\code{\link{bunddev_call_tidy}()},
\code{\link{bunddev_endpoints}()},
\code{\link{bunddev_ms_to_posix}()},
\code{\link{bunddev_parameter_values}()},
\code{\link{bunddev_parameters}()},
\code{\link{bunddev_parameters_for}()},
\code{\link{bunddev_spec}()},
\code{\link{bunddev_spec_path}()}
}
\concept{OpenAPI}
