% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_diga.R
\name{diga_organizations}
\alias{diga_organizations}
\title{List DiGA manufacturers}
\usage{
diga_organizations(
  params = list(),
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns.}
}
\value{
A tibble containing DiGA manufacturer organization resources from
the FHIR API. Structure depends on \code{flatten} setting.
}
\description{
List DiGA manufacturers
}
