% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_diga.R
\name{diga_questionnaires}
\alias{diga_questionnaires}
\title{List DiGA questionnaires}
\usage{
diga_questionnaires(
  params = list(),
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with FHIR entries.

Use \code{bunddev_parameters("diga")} to see the currently valid parameters if
the API has changed.
}
\description{
List DiGA questionnaires
}
\details{
Returns questionnaires from the DiGA directory.
Requires \code{DIGA_BEARER_TOKEN}.
}
\examples{
\dontrun{
diga_questionnaires()
}

}
