% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dip_bundestag.R
\name{dip_bundestag_vorgang}
\alias{dip_bundestag_vorgang}
\title{Get a DIP Vorgang}
\usage{
dip_bundestag_vorgang(
  vorgang_id,
  params = list(),
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{vorgang_id}{Vorgang id.}

\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with Vorgang metadata.
}
\description{
Get a DIP Vorgang
}
\details{
Returns metadata for a single Vorgang.
}
\examples{
\dontrun{
dip_bundestag_vorgang(84343)
}

}
