% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dwd.R
\name{dwd_avalanche_warnings}
\alias{dwd_avalanche_warnings}
\title{Fetch DWD avalanche warnings}
\usage{
dwd_avalanche_warnings(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with avalanche data.
}
\description{
Fetch DWD avalanche warnings
}
