% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_dwd.R
\name{dwd_sea_warning_text}
\alias{dwd_sea_warning_text}
\title{Fetch DWD sea warning text}
\usage{
dwd_sea_warning_text(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with the warning text.

Use \code{bunddev_parameters("dwd")} to see the currently valid parameters
if the API has changed.
}
\description{
Fetch DWD sea warning text
}
