% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_hilfsmittel.R
\name{hilfsmittel_produktart}
\alias{hilfsmittel_produktart}
\title{Get Hilfsmittel product type details}
\usage{
hilfsmittel_produktart(
  id,
  safe = TRUE,
  refresh = FALSE,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{id}{Produktart id.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with product type details.
}
\description{
Get Hilfsmittel product type details
}
\details{
Returns details for a product type (Produktart).
Official docs: https://github.com/bundesAPI/hilfsmittel-api.
}
\examples{
\dontrun{
hilfsmittel_produktart("e6b913ef-cf21-4c5f-826d-f866516c3c65")
}

}
