% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_hochwasserzentralen.R
\name{hochwasserzentralen_bundesland_info}
\alias{hochwasserzentralen_bundesland_info}
\title{Get flood portal metadata for a Bundesland}
\usage{
hochwasserzentralen_bundesland_info(
  bundesland_id,
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{bundesland_id}{Bundesland id (e.g., "HE").}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with Bundesland metadata.
}
\description{
Get flood portal metadata for a Bundesland
}
\details{
Returns metadata for a single Bundesland or region in the
hochwasserzentralen.de portal. Official docs:
https://bundesapi.github.io/hochwasserzentralen-api/.
}
\examples{
\dontrun{
hochwasserzentralen_bundesland_info("HE")
}

}
