% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_luftqualitaet.R
\name{luftqualitaet_networks}
\alias{luftqualitaet_networks}
\title{List networks}
\usage{
luftqualitaet_networks(params = list(), safe = TRUE, refresh = FALSE)
}
\arguments{
\item{params}{Query parameters.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with network metadata.

Use \code{bunddev_parameters("luftqualitaet")} to see the currently valid parameters
if the API has changed.
}
\description{
List networks
}
\details{
Returns networks used by the Luftqualitaet API. Official docs:
https://luftqualitaet.api.bund.dev.
}
