% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_nina.R
\name{nina_mowas_rss}
\alias{nina_mowas_rss}
\title{Get MOWAS RSS feed}
\usage{
nina_mowas_rss(ars, safe = TRUE, refresh = FALSE)
}
\arguments{
\item{ars}{ARS code.}

\item{safe}{Logical; if \code{TRUE} (default), apply rate-limiting and cache
GET responses to \code{tools::R_user_dir("bunddev", "cache")}.}

\item{refresh}{Logical; if \code{TRUE}, ignore cached responses and re-fetch
from the API (default \code{FALSE}).}
}
\value{
A one-row tibble with RSS XML text in column \code{xml}.
}
\description{
Get MOWAS RSS feed
}
\seealso{
Other NINA: 
\code{\link{nina_archive_mowas}()},
\code{\link{nina_archive_mowas_mapping}()},
\code{\link{nina_covid_infos}()},
\code{\link{nina_covid_map}()},
\code{\link{nina_covid_rules}()},
\code{\link{nina_covid_ticker}()},
\code{\link{nina_covid_ticker_message}()},
\code{\link{nina_dashboard}()},
\code{\link{nina_event_code}()},
\code{\link{nina_event_codes}()},
\code{\link{nina_faqs}()},
\code{\link{nina_logo}()},
\code{\link{nina_logos}()},
\code{\link{nina_mapdata}()},
\code{\link{nina_notfalltipps}()},
\code{\link{nina_version}()},
\code{\link{nina_warning}()},
\code{\link{nina_warning_geojson}()},
\code{\link{nina_warning_json}()},
\code{\link{nina_warnings}()}
}
\concept{NINA}
