% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_smard.R
\name{smard_indices}
\alias{smard_indices}
\title{List available SMARD timestamps}
\usage{
smard_indices(
  filter,
  region = "DE",
  resolution = "hour",
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{filter}{Filter id.}

\item{region}{Region code.}

\item{resolution}{Data resolution.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble of timestamps.

Use \code{bunddev_parameters("smard")} to see the currently valid parameters
if the API has changed.
}
\description{
List available SMARD timestamps
}
