% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_smard.R
\name{smard_timeseries}
\alias{smard_timeseries}
\title{Fetch SMARD timeseries data}
\usage{
smard_timeseries(
  filter,
  region = "DE",
  resolution = "hour",
  timestamp,
  safe = TRUE,
  refresh = FALSE
)
}
\arguments{
\item{filter}{Filter id.}

\item{region}{Region code.}

\item{resolution}{Data resolution.}

\item{timestamp}{Timestamp from indices.}

\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble of time series values.

Use \code{bunddev_parameters("smard")} to see the currently valid parameters
if the API has changed.
}
\description{
Fetch SMARD timeseries data
}
