% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_tagesschau.R
\name{tagesschau_channels}
\alias{tagesschau_channels}
\title{Fetch Tagesschau channels}
\usage{
tagesschau_channels(flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with channels.

Includes \code{date_time} as POSIXct in Europe/Berlin.
}
\description{
Fetch Tagesschau channels
}
\details{
Lists the Tagesschau channels endpoint. Official docs:
https://bundesapi.github.io/tagesschau-api/.
}
\examples{
\dontrun{
tagesschau_channels(flatten = TRUE)
}

}
\seealso{
\code{\link[=tagesschau_news]{tagesschau_news()}} and \code{\link[=tagesschau_homepage]{tagesschau_homepage()}}.
}
