% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_tagesschau.R
\name{tagesschau_homepage}
\alias{tagesschau_homepage}
\title{Fetch Tagesschau homepage items}
\usage{
tagesschau_homepage(flatten = FALSE, flatten_mode = "json")
}
\arguments{
\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with homepage items.

Includes \code{date_time} as POSIXct in Europe/Berlin.

Use \code{bunddev_parameters("tagesschau")} to see the currently valid
parameters if the API has changed.
}
\description{
Fetch Tagesschau homepage items
}
