% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_tagesschau.R
\name{tagesschau_news}
\alias{tagesschau_news}
\title{Fetch Tagesschau news items}
\usage{
tagesschau_news(
  regions = NULL,
  ressort = NULL,
  flatten = FALSE,
  flatten_mode = "json"
)
}
\arguments{
\item{regions}{Optional region ids.}

\item{ressort}{Optional ressort filter.}

\item{flatten}{Logical; drop nested list columns.}

\item{flatten_mode}{Flatten strategy for list columns. Use "unnest" to
expand list-columns into multiple rows.}
}
\value{
A tibble with news items.
}
\description{
Fetch Tagesschau news items
}
