% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_travelwarning.R
\name{travelwarning_representatives_germany}
\alias{travelwarning_representatives_germany}
\title{List foreign representatives in Germany}
\usage{
travelwarning_representatives_germany(safe = TRUE, refresh = FALSE)
}
\arguments{
\item{safe}{Logical; apply throttling and caching.}

\item{refresh}{Logical; refresh cached responses.}
}
\value{
A tibble with representatives.

Includes \code{last_modified_time} as POSIXct in Europe/Berlin.
}
\description{
List foreign representatives in Germany
}
\details{
Returns a list of foreign representatives in Germany. Official docs:
https://travelwarning.api.bund.dev.
}
\examples{
\dontrun{
travelwarning_representatives_germany()
}

}
